Moves globalStyles file to common directory and updates paths
